Move models out of myapp.py.
